requirements.dev.txt
flake8>=3.9.2,<3.10
django-environ==0.11.2